wip feat: filter init
