+++ console/packages/starwhale-ui/src/Search/Filter.tsx
@@ -1,0 +1,4 @@
+function Filter() {
+    return {}
+}
+export default Filter
