+++ console/packages/starwhale-ui/src/Search/FilterString.tsx
@@ -1,0 +1,2 @@
+function FilterString() {}
+export default FilterString
