+++ console/packages/starwhale-ui/src/Search/index.ts
@@ -1,1 +1,1 @@
-export { default } from './Search'
+export { default } from './SearchComponent'
